pyproject.toml
[project]
name = "logisty-back"
version = "0.1.0"
description = ""
authors = [
    { name = "Andrea Rabe", email = "andrearaberabe@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "djoser (>=2.3.1,<3.0.0)",
    "social-auth-app-django (>=5.4.2,<6.0.0)",
    "psycopg (>=3.2.4,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "pytest-django (>=4.9.0,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "faker (>=35.2.0,<36.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
factory-boy = "^3.3.3"
faker = "^35.2.0"

